Rename Connections Component to StopConnections
