+++ stops/apps/frontend/src/components/Stop/SpecificStop/index.tsx
@@ -1,24 +1,17 @@
 'use client';
 
-/* * */
-
 import type { Stop } from '@tmlmobilidade/types';
 
-/* * */
-
 import { useStopDetailContext } from '@/contexts/StopDetail.context';
-// import { getUnixTimestamp } from '@tmlmobilidade/utils';
-
-import { getUnixTimestampFromJSDate } from '@tmlmobilidade/utils';
 
 import Comments from '../Comments';
-import Connections from '../Connections';
 import MapContainer from '../MapContainer';
 import Media from '../Media';
 import Observations from '../Observations';
 import StopAccessibility from '../StopAccessibility';
 import StopAdminInformation from '../StopAdminInformation';
 import StopAffectation from '../StopAffectation';
+import StopConnections from '../StopConnections';
 import StopDetails from '../StopDetails';
 import StopEquipments from '../StopEquipments';
 import StopInfrasctructure from '../StopInfrasctructure';
@@ -28,31 +21,11 @@ import styles from '../styles.module.css';
 
 /* * */
 
-// export default function Stop({ stop }: SpecificStopProps) {
 export default function Stop() {
 	//
 
 	//
-	// A. Setup variables
-
-	const stopDetailContext = useStopDetailContext();
-
-	const { data: stop } = stopDetailContext;
-	console.log('=> stop', stop);
-
-	//
-	// B. Transform data
-
-	const last_schedules_check_setter = (date: Date) => {
-		stop.form.setFieldValue('last_schedules_check', getUnixTimestampFromJSDate(date));
-	};
-
-	const last_schedules_maintenance_setter = (date: Date) => {
-		stop.form.setFieldValue('last_schedules_maintenance', getUnixTimestampFromJSDate(date));
-	};
-
-	//
-	// C. Render components
+	// A. Render components
 
 	return (
 		<div className={styles.container}>
@@ -65,7 +38,7 @@ export default function Stop() {
 			<StopPublicInformation />
 			<StopAccessibility />
 			<StopEquipments />
-			<Connections />
+			<StopConnections />
 			{/*
 			<Media
 				file_ids={stop.form.getValues().file_ids || ['']}
